通信业务实现
