+++ Ifx-Server/pom.xml
@@ -10,7 +10,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>Ifx-Server</artifactId>
-<!--    <packaging>pom</packaging>-->
+    <packaging>pom</packaging>
+    <modules>
+        <module>server-api</module>
+        <module>server-service</module>
+    </modules>
+    <!--    <packaging>pom</packaging>-->
 
     <properties>
 <!--        <maven.compiler.source>11</maven.compiler.source>-->
