+++ Ifx-Server/server-api/src/main/java/com/ifx/server/service/ClientSendService.java
@@ -1,0 +1,13 @@
+package com.ifx.server.service;
+
+public interface ClientSendService {
+
+    /**
+     * 发送消息至客户端
+     * @param accountId
+     * @param msg
+     */
+    public void sendClient(String accountId,String msg );
+
+
+}
