+++ Ifx-Server/server-service/src/main/java/com/ifx/server/invoke/GateInvoke.java
@@ -1,0 +1,16 @@
+package com.ifx.server.invoke;
+
+import com.ifx.connect.proto.Protocol;
+import io.netty.channel.ChannelHandlerContext;
+
+public interface GateInvoke {
+
+
+    public void doWork(ChannelHandlerContext channel, Protocol protocol);
+
+    public void doException(Throwable e);
+
+
+
+
+}
