+++ Ifx-Server/server-service/src/main/java/com/ifx/server/netty/holder/NettyContext.java
@@ -1,0 +1,21 @@
+package com.ifx.server.netty.holder;
+
+import io.netty.channel.Channel;
+
+import java.util.concurrent.ConcurrentHashMap;
+
+//@Component
+public class NettyContext {
+    private static ConcurrentHashMap<String, Channel> channelMap ;
+
+    public void addAccount(Channel channel ,String account){
+        if (channelMap == null ){
+            channelMap = new ConcurrentHashMap<>();
+        }
+        channelMap.put(account,channel);
+    }
+
+    public Channel getChannelByAccount(String account){
+        return channelMap.computeIfAbsent(account, (k)-> null);
+    }
+}
