+++ Ifx-Server/server-service/src/main/java/com/ifx/server/route/ServerRoute.java
@@ -1,0 +1,13 @@
+package com.ifx.server.route;
+
+import com.ifx.connect.proto.Protocol;
+
+
+public interface ServerRoute {
+
+    public void getMeteService();
+
+    public void toService(Protocol protocol);
+
+
+}
