--- Ifx-Server/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,1 +1,0 @@
-AccountFilter=com.ifx.server.filter.AccountFilter
