+++ ifx-connect/src/main/java/com/ifx/connect/proto/dubbo/DubboApiConst.java
@@ -1,6 +1,11 @@
 package com.ifx.connect.proto.dubbo;
 
+/**
+ * 基于串协议的实现
+ * 此版本不做拓展
+ */
 public interface DubboApiConst {
+
     String signature = "interface";
     String method ="method";
