+++ ifx-fxClient/src/main/resource/application.yaml
@@ -1,0 +1,6 @@
+ifx:
+  connect:
+    client:
+      serverHost: 127.0.0.1
+      serverPort: 8099
+      connectTimeOut: 60000
